refactor(mapgen): remove mechanical foundation lib owners
