+++ mods/mod-swooper-maps/src/domain/foundation/model/policy/reference-area.ts
