+++ mods/mod-swooper-maps/src/domain/foundation/model/policy/tectonic-event-types.ts
@@ -1,0 +1,7 @@
+export const EVENT_TYPE = {
+  convergenceSubduction: 1,
+  convergenceCollision: 2,
+  divergenceRift: 3,
+  transformShear: 4,
+  intraplateHotspot: 5,
+} as const;
