+++ mods/mod-swooper-maps/src/domain/foundation/ops/compute-plate-graph/index.ts
@@ -2,7 +2,10 @@ import { createLabelRng } from "@swooper/mapgen-core";
 import { createOp, type Static } from "@swooper/mapgen-core/authoring";
 import { clamp01, wrapDeltaPeriodic } from "@swooper/mapgen-core/lib/math";
 
-import { deriveFoundationReferenceArea, requireEnvDimensions } from "../../lib/normalize.js";
+import {
+  deriveFoundationReferenceArea,
+  requireEnvDimensions,
+} from "../../model/policy/reference-area.js";
 import { requireCrust, requireMesh } from "../../lib/require.js";
 import type { FoundationPlate } from "./contract.js";
 import ComputePlateGraphContract from "./contract.js";
