+++ mods/mod-swooper-maps/src/domain/foundation/ops/compute-tectonic-provenance/rules/index.ts
@@ -3,18 +3,19 @@ import {
   requireMesh as requireMeshInput,
   requirePlateGraph as requirePlateGraphInput,
 } from "../../../lib/require.js";
+import { EVENT_TYPE } from "../../../model/policy/tectonic-event-types.js";
 import {
   ARC_RESET_THRESHOLD_FRAC_OF_MAX,
   ARC_RESET_THRESHOLD_MIN,
-  EVENT_TYPE,
   HOTSPOT_RESET_THRESHOLD_FRAC_OF_MAX,
   HOTSPOT_RESET_THRESHOLD_MIN,
   RIFT_RESET_THRESHOLD_FRAC_OF_MAX,
   RIFT_RESET_THRESHOLD_MIN,
-} from "../../../lib/tectonics/constants.js";
+  deriveResetThreshold,
+} from "./reset-threshold-policy.js";
 import type { FoundationTectonicEraFieldsInternal } from "../../../lib/tectonics/internal-contract.js";
 import type { FoundationTectonicProvenance } from "../../../lib/tectonics/schemas.js";
-import { clampByte, deriveResetThreshold } from "../../../lib/tectonics/shared.js";
+import { clampByte } from "../../../lib/tectonics/shared.js";
 import type { FoundationMesh } from "../../compute-mesh/contract.js";
 import type { FoundationPlateGraph } from "../../compute-plate-graph/contract.js";
