+++ mods/mod-swooper-maps/src/domain/foundation/ops/compute-tracer-advection/rules/constants.ts
@@ -1,0 +1,1 @@
+export const ADVECTION_STEPS_PER_ERA = 6;
