+++ mods/mod-swooper-maps/src/domain/foundation/ops/compute-tracer-advection/rules/index.ts
@@ -2,7 +2,7 @@ import {
   requireMantleForcing as requireMantleForcingInput,
   requireMesh as requireMeshInput,
 } from "../../../lib/require.js";
-import { ADVECTION_STEPS_PER_ERA } from "../../../lib/tectonics/constants.js";
+import { ADVECTION_STEPS_PER_ERA } from "./constants.js";
 import type { FoundationTectonicEraFieldsInternal } from "../../../lib/tectonics/internal-contract.js";
 import { chooseDriftNeighbor, normalizeToInt8 } from "../../../lib/tectonics/shared.js";
 import type { FoundationMantleForcing } from "../../compute-mantle-forcing/contract.js";
